Always ensure shell file is set up for chruby
If the script failed for some reason, and then chruby was installed manually, when the script was run again, it would not automatically update the shell file. This fixes that scenario so that even if chruby is already installed, the script still checks to see if the shell file has the proper commands, and adds them if it doesn’t.

I also removed the brew update line because Homebrew now automatically updates when you run brew bundle or brew install.