Specify OpenSSL directory when installing ruby
Prior to version 0.8.0, `ruby-install` would automatically set the OpenSSL to the one installed by Homebrew. For some reason, this path is no longer set in 0.8.0, which was causing ruby installations to fail.

By searching for “openssl” in the `ruby-install` repo, I came across the `--with-openssl-dir` flag, and that seems to fix the issue. I tried it on my computer and it worked fine. Hopefully it works for others too.