Add Nodenv and yarn
**Why**: To make it easier for folks to create Rails apps with webpacker.

I also removed `chruby-fish` because it's buggy. I added a note to the
Readme with a workaround for those who really want to use fish, like me.

Finally, I added an option to skip `.laptop.local` if you care mostly
about updating the main part of the script. In my case, I have a lot of
Mac Apps in my Brewfile.local, and sometimes I just want to test the
script and not have to wait for Homebrew to go through all the apps.